chore: separate AnimatorInstaller
