+++ Editor/Animator/AnimatorInstaller.cs
@@ -1,96 +1,103 @@
 using UnityEditor.Animations;
-using nadena.dev.ndmf;
 using nadena.dev.ndmf.animator;
-using nadena.dev.modular_avatar.core;
-using aoyon.facetune.platform;
 using aoyon.facetune.build;
-using VRC.SDKBase;
-using VRC.SDK3.Avatars.Components;
 
 namespace aoyon.facetune.animator;
 
-internal class AnimatorInstaller
+internal class AnimatorInstaller : InstallerBase
 {
-    private readonly VirtualAnimatorController _virtualController;
-
-    private readonly SessionContext _sessionContext;
-    private readonly IPlatformSupport _platformSupport;
-
-    private readonly bool _useWriteDefaults;
     private readonly float _transitionDurationSeconds;
 
     private readonly Dictionary<Expression, VirtualClip> _expressionClipCache = new();
-    private readonly Dictionary<AdvancedEyeBlinkSettings, int> _advancedEyeBlinkIndex = new();
-    private readonly Dictionary<AdvancedLipSyncSettings, int> _advancedLipSyncIndex = new();
-    private readonly VirtualClip _emptyClip;
 
-    private static readonly Vector3 EntryStatePosition = new Vector3(50, 120, 0);
-    private static readonly Vector3 FTDefaultStatePosition = new Vector3(300, 0, 0);
-    private const float PositionXStep = 250;
-    private const float PositionYStep = 50;
-    
+    private readonly LipSyncInstaller _lipSyncInstaller;
+    private readonly BlinkInstaller _blinkInstaller;
+
     private const int InitLayerPriority = -1; // 上書きを意図しない初期化レイヤー。
-    private const int LayerPriority = 1; // FaceEmo: 0
-
-    private const string TrueParameterName = $"{FaceTuneConsts.ParameterPrefix}/True";
-    // LipSync
-    private const string LipSyncParameterPrefix = $"{FaceTuneConsts.ParameterPrefix}/LipSync";
-    private const string LipSyncAllowAAP = $"{LipSyncParameterPrefix}/Allow";
-    private const string LipSyncUseAdvancedAAP = $"{LipSyncParameterPrefix}/UseAdvanced";
-    private const string LipSyncUseCancelerAAP = $"{LipSyncParameterPrefix}/UseCanceler";
-    private const string LipSyncModeAAP = $"{LipSyncParameterPrefix}/Mode";
-
-    // Blink
-    private const string BlinkParameterPrefix = $"{FaceTuneConsts.ParameterPrefix}/Blink";
-    private const string BlinkAllowAAP = $"{BlinkParameterPrefix}/Allow";
-    private const string BlinkUseAnimationAAP = $"{BlinkParameterPrefix}/UseAnimation";
-    private const string BlinkAnimationModeAAP = $"{BlinkParameterPrefix}/AnimationMode";
-    private const string BlinkUseCancelerAAP = $"{BlinkParameterPrefix}/UseCanceler";
-    private const string BlinkBlinkingAAP = $"{BlinkParameterPrefix}/Blinking";
-    private const string BlinkDelayMultiplier = $"{BlinkParameterPrefix}/DelayMultiplier";
-
-    public AnimatorInstaller(SessionContext sessionContext, VirtualAnimatorController virtualController, bool useWriteDefaults)
+    private static readonly Vector3 ExclusiveStatePosition = new Vector3(300, 0, 0);
+    
+    public AnimatorInstaller(VirtualAnimatorController virtualController, SessionContext sessionContext, bool useWriteDefaults) : base(virtualController, sessionContext, useWriteDefaults)
     {
-        _virtualController = virtualController;
-
-        _sessionContext = sessionContext;
-        _platformSupport = platform.PlatformSupport.GetSupport(_sessionContext.Root.transform);
-
         _transitionDurationSeconds = 0.1f; // 変更可能にすべき？
-        _useWriteDefaults = useWriteDefaults;
-
-        _emptyClip = AnimatorHelper.CreateCustomEmpty();
-
-        // Todo: 必要なパラメーターのみの追加
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Bool, TrueParameterName).defaultBool = true;
-
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, LipSyncAllowAAP);
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, LipSyncUseCancelerAAP);
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, LipSyncModeAAP);
-
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, BlinkAllowAAP);
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, BlinkUseAnimationAAP);
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, BlinkAnimationModeAAP);
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, BlinkUseCancelerAAP);
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, BlinkBlinkingAAP).defaultFloat = -1f; // -1: Stare, 0: Closing, 1: Opening
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, BlinkDelayMultiplier);
-    }
-
-    private static VirtualLayer AddFTLayer(VirtualAnimatorController controller, string layerName, int priority)
-    {
-        var layerPriority = new LayerPriority(priority);
-        var layer = controller.AddLayer(layerPriority, $"{FaceTuneConsts.ShortName}: {layerName}");
-        layer.StateMachine!.EnsureBehavior<ModularAvatarMMDLayerControl>().DisableInMMDMode = true;
-        return layer; 
-    }
-
-    private VirtualState AddFTState(VirtualLayer layer, string stateName, Vector3? position = null)
-    {
-        var state = layer.StateMachine!.AddState(stateName, position: position);
-        state.WriteDefaultValues = _useWriteDefaults;
-        return state;
+        _lipSyncInstaller = new LipSyncInstaller(virtualController, sessionContext, useWriteDefaults);
+        _blinkInstaller = new BlinkInstaller(virtualController, sessionContext, useWriteDefaults);
     }
     
+    public void DisableExistingControlAndInstallPatternData(InstallData installData)
+    {
+        var patternData = installData.PatternData;
+        if (patternData.IsEmpty) return;
+
+        CreateDefaultLayer(installData.OverrideBlendShapes, installData.OverrideProperties, patternData);
+
+        foreach (var patternGroup in patternData.GetConsecutiveTypeGroups())
+        {
+            var type = patternGroup.Type;
+            if (type == typeof(Preset))
+            {
+                var presets = patternGroup.Group.Select(item => (Preset)item).ToList();
+                InstallPresetGroup(presets, LayerPriority);
+            }
+            else if (type == typeof(SingleExpressionPattern))
+            {
+                var singleExpressionPatterns = patternGroup.Group.Select(item => (SingleExpressionPattern)item).ToList();
+                InstallSingleExpressionPatternGroup(singleExpressionPatterns, LayerPriority);
+            }
+        }
+        
+        var patternExpressions = patternData.GetAllExpressions();
+        if (patternExpressions.Any(e => e.FacialSettings.AllowEyeBlink == TrackingPermission.Disallow
+            || e.FacialSettings.AdvancedEyBlinkSettings.UseAdvancedEyeBlink))
+        {
+            _blinkInstaller.AddEyeBlinkLayer();
+        }
+        if (patternExpressions.Any(e => e.FacialSettings.AllowLipSync == TrackingPermission.Disallow
+            || e.FacialSettings.AdvancedLipSyncSettings.IsEnabled()))
+        {
+            _lipSyncInstaller.AddLipSyncLayer();
+        }
+    }
+
+    // ブレンドシェイプ及びアニメーション用の初期化レイヤーを生成する
+    // DisbaleExisingControlが無効な場合は既存の制御より低い優先度、有効な場合は高い優先度
+    // 優先度が同じ場合はブレンドシェイプ用とアニメーション用のレイヤーを結合する(軽量化)。
+    private void CreateDefaultLayer(bool overrideShapes, bool overrideProperties, PatternData patternData)
+    {
+        VirtualLayer defaultLayer;
+        VirtualState defaultState;
+
+        var shapesLayerPriority = overrideShapes ? LayerPriority : InitLayerPriority;
+        var propertiesLayerPriority = overrideProperties ? LayerPriority : InitLayerPriority;
+
+        // ブレンドシェイプの初期化レイヤー
+        var shapesLayer = AddFTLayer(_virtualController, "Default", shapesLayerPriority);
+        var shapesState = AddFTState(shapesLayer, "Default", ExclusiveStatePosition);
+        AddAnimationToState(shapesState, _sessionContext.ZeroWeightBlendShapes.ToGenericAnimations(_sessionContext.BodyPath));
+
+        var bindings = patternData.GetAllExpressions().SelectMany(e => e.Animations).Select(a => a.CurveBinding).Distinct().ToList();
+
+        var facialBinding = SerializableCurveBinding.FloatCurve(_sessionContext.BodyPath, typeof(SkinnedMeshRenderer), FaceTuneConsts.AnimatedBlendShapePrefix);
+        var nonFacialBindings = bindings.Where(b => b != facialBinding);
+        if (!nonFacialBindings.Any()) return;
+
+        var defaultPropertiesAnimations = AnimatorHelper.GetDefaultValueAnimations(_sessionContext.Root, nonFacialBindings);
+        if (!defaultPropertiesAnimations.Any()) return;
+
+        if (shapesLayerPriority == propertiesLayerPriority)
+        {
+            AddAnimationToState(shapesState, defaultPropertiesAnimations);
+        }
+        else
+        {
+            var propertiesLayer = AddFTLayer(_virtualController, "Default", propertiesLayerPriority);
+            var propertiesState = AddFTState(propertiesLayer, "Default", ExclusiveStatePosition);
+            AddAnimationToState(propertiesState, defaultPropertiesAnimations);
+        }
+
+        defaultLayer = shapesLayer;
+        defaultState = shapesState;
+    }
+
     private void AddExpressionToState(VirtualState state, Expression expression)
     {
         if (state.TryGetClip(out var clip))
@@ -137,106 +144,6 @@ internal class AnimatorInstaller
         }
     }
 
-    private void AddAnimationToState(VirtualState state, IEnumerable<GenericAnimation> animations)
-    {
-        var clip = state.GetOrCreateClip(state.Name);
-        AddAnimationToState(clip, animations);
-    }
-
-    private void AddAnimationToState(VirtualClip clip, IEnumerable<GenericAnimation> animations)
-    {
-        clip.SetAnimations(animations);
-    }
-
-    // ブレンドシェイプ及びアニメーション用の初期化レイヤーを生成する
-    // DisbaleExisingControlが無効な場合は既存の制御より低い優先度、有効な場合は高い優先度
-    // 優先度が同じ場合はブレンドシェイプ用とアニメーション用のレイヤーを結合する(軽量化)。
-    private void CreateDefaultLayer(bool overrideShapes, bool overrideProperties, PatternData patternData)
-    {
-        VirtualLayer defaultLayer;
-        VirtualState defaultState;
-
-        var shapesLayerPriority = overrideShapes ? LayerPriority : InitLayerPriority;
-        var propertiesLayerPriority = overrideProperties ? LayerPriority : InitLayerPriority;
-
-        // ブレンドシェイプの初期化レイヤー
-        var shapesLayer = AddFTLayer(_virtualController, "Default", shapesLayerPriority);
-        var shapesState = AddFTState(shapesLayer, "Default", FTDefaultStatePosition);
-        AddAnimationToState(shapesState, _sessionContext.ZeroWeightBlendShapes.ToGenericAnimations(_sessionContext.BodyPath));
-
-        var bindings = patternData.GetAllExpressions().SelectMany(e => e.Animations).Select(a => a.CurveBinding).Distinct().ToList();
-
-        var facialBinding = SerializableCurveBinding.FloatCurve(_sessionContext.BodyPath, typeof(SkinnedMeshRenderer), FaceTuneConsts.AnimatedBlendShapePrefix);
-        var nonFacialBindings = bindings.Where(b => b != facialBinding);
-        if (!nonFacialBindings.Any()) return;
-
-        var defaultPropertiesAnimations = AnimatorHelper.GetDefaultValueAnimations(_sessionContext.Root, nonFacialBindings);
-        if (!defaultPropertiesAnimations.Any()) return;
-
-        if (shapesLayerPriority == propertiesLayerPriority)
-        {
-            AddAnimationToState(shapesState, defaultPropertiesAnimations);
-        }
-        else
-        {
-            var propertiesLayer = AddFTLayer(_virtualController, "Default", propertiesLayerPriority);
-            var propertiesState = AddFTState(propertiesLayer, "Default", FTDefaultStatePosition);
-            AddAnimationToState(propertiesState, defaultPropertiesAnimations);
-        }
-
-        defaultLayer = shapesLayer;
-        defaultState = shapesState;
-    }
-
-    public void DisableExistingControlAndInstallPatternData(InstallData installData)
-    {
-        var patternData = installData.PatternData;
-        if (patternData.IsEmpty) return;
-
-        CreateDefaultLayer(installData.OverrideBlendShapes, installData.OverrideProperties, patternData);
-
-
-        foreach (var patternGroup in patternData.GetConsecutiveTypeGroups())
-        {
-            var type = patternGroup.Type;
-            if (type == typeof(Preset))
-            {
-                var presets = patternGroup.Group.Select(item => (Preset)item).ToList();
-                InstallPresetGroup(presets, LayerPriority);
-            }
-            else if (type == typeof(SingleExpressionPattern))
-            {
-                var singleExpressionPatterns = patternGroup.Group.Select(item => (SingleExpressionPattern)item).ToList();
-                InstallSingleExpressionPatternGroup(singleExpressionPatterns, LayerPriority);
-            }
-        }
-        
-        var patternExpressions = patternData.GetAllExpressions();
-        if (patternExpressions.Any(e => e.FacialSettings.AllowEyeBlink == TrackingPermission.Disallow
-            || e.FacialSettings.AdvancedEyBlinkSettings.UseAdvancedEyeBlink))
-        {
-            AddEyeBlinkLayer();
-        }
-        if (patternExpressions.Any(e => e.FacialSettings.AllowLipSync == TrackingPermission.Disallow
-            || e.FacialSettings.AdvancedLipSyncSettings.IsEnabled()))
-        {
-            AddLipSyncLayer();
-        }
-    }
-
-    private void AsPassThrough(VirtualState state)
-    {
-        // Transition Durationを用いて上のレイヤーとブレンドする際、WD OFFの場合は空のClip、WD ONの場合はNone
-        if (_useWriteDefaults)
-        {
-            state.Motion = null;
-        }
-        else
-        {
-            state.Motion = _emptyClip;
-        }
-    }
-
     private void InstallPresetGroup(IReadOnlyList<Preset> presets, int priority)
     {
         var maxPatterns = presets.Max(p => p.Patterns.Count); 
@@ -257,7 +164,7 @@ internal class AnimatorInstaller
                 if (!layerCreatedForThisIndex)
                 {
                     layers[i] = AddFTLayer(_virtualController, $"Preset Pattern Group {i}", priority); 
-                    defaultStates[i] = AddFTState(layers[i], "PassThrough", FTDefaultStatePosition);
+                    defaultStates[i] = AddFTState(layers[i], "PassThrough", ExclusiveStatePosition);
                     AsPassThrough(defaultStates[i]);
                     layerCreatedForThisIndex = true;
                 }
@@ -277,9 +184,9 @@ internal class AnimatorInstaller
                 !singleExpressionPattern.ExpressionPattern.ExpressionWithConditions.Any()) continue;
 
             var layer = AddFTLayer(_virtualController, singleExpressionPattern.Name, priority);
-            var defaultState = AddFTState(layer, "PassThrough", FTDefaultStatePosition);
+            var defaultState = AddFTState(layer, "PassThrough", ExclusiveStatePosition);
             AsPassThrough(defaultState);
-            var basePosition = FTDefaultStatePosition + new Vector3(0, 2 * PositionYStep, 0);
+            var basePosition = ExclusiveStatePosition + new Vector3(0, 2 * PositionYStep, 0);
             AddExpressionWithConditions(layer, defaultState, singleExpressionPattern.ExpressionPattern.ExpressionWithConditions, basePosition); 
         }
     }
@@ -376,610 +283,7 @@ internal class AnimatorInstaller
     private void SetFacialSettings(VirtualClip clip, FacialSettings? facialSettings)
     {
         if (facialSettings == null) return;
-
-        // EyeBlink
-        if (facialSettings.AllowEyeBlink != TrackingPermission.Keep)
-        {
-            // Allow
-            var allowBlinkCurve = new AnimationCurve();
-            var value = facialSettings.AllowEyeBlink == TrackingPermission.Allow ? 1 : 0;
-            allowBlinkCurve.AddKey(0, value);
-            clip.SetFloatCurve("", typeof(Animator), BlinkAllowAAP, allowBlinkCurve);
-
-            var advancedSettings = facialSettings.AdvancedEyBlinkSettings;
-            if (advancedSettings.UseAdvancedEyeBlink && advancedSettings.UseAnimation)
-            {
-                // UseAnimation
-                var useAnimationCurve = new AnimationCurve();
-                useAnimationCurve.AddKey(0, 1);
-                clip.SetFloatCurve("", typeof(Animator), BlinkUseAnimationAAP, useAnimationCurve);
-
-                // AnimationMode
-                var index = GetAdvancedEyeBlinkIndex(advancedSettings);
-                var animationModeCurve = new AnimationCurve();
-                animationModeCurve.AddKey(0, VRCAAPHelper.IndexToValue(index));
-                clip.SetFloatCurve("", typeof(Animator), BlinkAnimationModeAAP, animationModeCurve);
-
-                // UseCanceler
-                if (advancedSettings.IsCancelerEnabled())
-                {
-                    var useCancelerCurve = new AnimationCurve();
-                    useCancelerCurve.AddKey(0, 1);
-                    clip.SetFloatCurve("", typeof(Animator), BlinkUseCancelerAAP, useCancelerCurve);
-                }
-            }
-        }
-
-        // LipSync
-        if (facialSettings.AllowLipSync != TrackingPermission.Keep)
-        {
-            var curve = new AnimationCurve();
-            var value = facialSettings.AllowLipSync == TrackingPermission.Allow ? 1 : 0;
-            curve.AddKey(0, value);
-            clip.SetFloatCurve("", typeof(Animator), LipSyncAllowAAP, curve);
-
-            var advancedSettings = facialSettings.AdvancedLipSyncSettings;
-            if (advancedSettings.IsEnabled())
-            {
-                var useAdvancedCurve = new AnimationCurve();
-                useAdvancedCurve.AddKey(0, 1);
-                clip.SetFloatCurve("", typeof(Animator), LipSyncUseAdvancedAAP, useAdvancedCurve);
-
-                var index = GetAdvancedLipSyncIndex(advancedSettings);
-                var modeCurve = new AnimationCurve();
-                modeCurve.AddKey(0, VRCAAPHelper.IndexToValue(index));
-                clip.SetFloatCurve("", typeof(Animator), LipSyncModeAAP, modeCurve);
-
-                if (advancedSettings.IsCancelerEnabled())
-                {
-                    var useCancelerCurve = new AnimationCurve();
-                    useCancelerCurve.AddKey(0, 1);
-                    clip.SetFloatCurve("", typeof(Animator), LipSyncUseCancelerAAP, useCancelerCurve);
-                }
-            }
-        }
-    }
-
-    private int GetAdvancedEyeBlinkIndex(AdvancedEyeBlinkSettings advancedEyeBlinkSettings)
-    {
-        if (_advancedEyeBlinkIndex.TryGetValue(advancedEyeBlinkSettings, out var index))
-        {
-            return index;
-        }
-        else
-        {
-            index = _advancedEyeBlinkIndex.Count;
-            _advancedEyeBlinkIndex[advancedEyeBlinkSettings] = index;
-        }
-        return index;
-    }
-
-    private int GetAdvancedLipSyncIndex(AdvancedLipSyncSettings advancedLipSyncSettings)
-    {
-        if (_advancedLipSyncIndex.TryGetValue(advancedLipSyncSettings, out var index))
-        {
-            return index;
-        }
-        else
-        {
-            index = _advancedLipSyncIndex.Count;
-            _advancedLipSyncIndex[advancedLipSyncSettings] = index;
-        }
-        return index;
-    }
-
-    private void AddEyeBlinkLayer()
-    {
-        VirtualLayer? cancelerLayer = null;
-        VirtualLayer? eyeBlinkLayer = null;
-        
-        if (_advancedEyeBlinkIndex.Keys.Any(k => k.IsCancelerEnabled()))
-        {
-            cancelerLayer = AddFTLayer(_virtualController, "EyeBlink (Canceler)", LayerPriority);
-        }
-        eyeBlinkLayer = AddFTLayer(_virtualController, "EyeBlink", LayerPriority);
-
-        // 最初のフレームでTraking Controlを変更すると巻き戻される。
-        // 2フレームの遅延が必要らしいので多めに0.1s程度遅延させる
-        var delayState = AddFTState(eyeBlinkLayer, "Delay", EntryStatePosition + new Vector3(-20, 2 * PositionYStep, 0));
-        var delayClip = AnimatorHelper.CreateDelayClip(0.1f);
-        delayState.Motion = delayClip;
-
-        // DefaultのTracking/Animationの入れ替え
-        var enabledPosition = EntryStatePosition + new Vector3(PositionXStep, 0, 0);
-        var enabled = AddFTState(eyeBlinkLayer, "Default: Enabled", enabledPosition);
-        var disabled = AddFTState(eyeBlinkLayer, "Default: Disabled", enabledPosition + new Vector3(0, 2 * PositionYStep, 0));
-        
-        enabled.Motion = _emptyClip;
-        disabled.Motion = _emptyClip;
-        _platformSupport.SetEyeBlinkTrack(enabled, true);
-        _platformSupport.SetEyeBlinkTrack(disabled, false);
-
-        // Delay -> Enabled
-        var delayToEnabled = AnimatorHelper.CreateTransitionWithExitTime(1f, 0f);
-        delayToEnabled.SetDestination(enabled);
-        delayState.Transitions = delayState.Transitions.Add(delayToEnabled);
-
-        // Enabled -> Disabled
-        var enabledToDisabled = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-        enabledToDisabled.SetDestination(disabled);
-        enabledToDisabled.Conditions = ImmutableList.Create(new AnimatorCondition()
-        {
-            parameter = BlinkAllowAAP,
-            mode = AnimatorConditionMode.Less,
-            threshold = 0.99f
-        });
-        enabled.Transitions = enabled.Transitions.Add(enabledToDisabled);
-
-        // Disabled -> Enabled
-        var disabledToEnabled = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-        disabledToEnabled.SetDestination(enabled);
-        disabledToEnabled.Conditions = ImmutableList.Create(new AnimatorCondition()
-        {
-            parameter = BlinkAllowAAP,
-            mode = AnimatorConditionMode.Greater,
-            threshold = 0.99f
-        });
-        disabled.Transitions = disabled.Transitions.Add(disabledToEnabled);
-
-        if (_advancedEyeBlinkIndex.Count == 0 || _advancedEyeBlinkIndex.Keys.All(k => !k.IsEnabled())) return;
-    
-        // AnimationGate
-        var disableTrackingPosition = enabledPosition + new Vector3(PositionXStep, 0, 0);
-        var disableTracking = AddFTState(eyeBlinkLayer, "DisableTracking", disableTrackingPosition);
-        var animationGatePosition = disableTrackingPosition + new Vector3(0, 2 * PositionYStep, 0);
-        var animationGate = AddFTState(eyeBlinkLayer, "AnimationGate", animationGatePosition);
-
-        disableTracking.Motion = _emptyClip;
-        animationGate.Motion = _emptyClip;
-        _platformSupport.SetEyeBlinkTrack(disableTracking, false);
-
-        // Disabled -> AnimationGate
-        var disabledToAnimationGate = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-        disabledToAnimationGate.SetDestination(animationGate);
-        disabledToAnimationGate.Conditions = ImmutableList.Create(new AnimatorCondition()
-        {
-            parameter = BlinkUseAnimationAAP,
-            mode = AnimatorConditionMode.Greater,
-            threshold = 0f
-        });
-        disabled.Transitions = disabled.Transitions.Add(disabledToAnimationGate);
-
-        // AnimationGate -> Disabled
-        var animationGateToDsiabled = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-        animationGateToDsiabled.SetDestination(disabled);
-        animationGateToDsiabled.Conditions = ImmutableList.Create(new AnimatorCondition()
-        {
-            parameter = BlinkUseAnimationAAP,
-            mode = AnimatorConditionMode.Less,
-            threshold = 1f
-        });
-        animationGate.Transitions = animationGate.Transitions.Add(animationGateToDsiabled);
-
-        // Enabled -> DisableTracking
-        var enabledToDisableTracking = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-        enabledToDisableTracking.SetDestination(disableTracking);
-        enabledToDisableTracking.Conditions = ImmutableList.Create(new AnimatorCondition()
-        {
-            parameter = BlinkUseAnimationAAP,
-            mode = AnimatorConditionMode.Greater,
-            threshold = 0f
-        });
-        enabled.Transitions = enabled.Transitions.Add(enabledToDisableTracking);
-
-        // DisableTracking -> AnimationGate
-        var disableTrackingToAnimationGate = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-        disableTrackingToAnimationGate.SetDestination(animationGate);
-        disableTrackingToAnimationGate.Conditions = ImmutableList.Create(new AnimatorCondition()
-        {
-            parameter = TrueParameterName,
-            mode = AnimatorConditionMode.If
-        });
-        disableTracking.Transitions = disableTracking.Transitions.Add(disableTrackingToAnimationGate);
-
-        // Animationを用いた瞬き制御
-
-        var closeShapeNames = _advancedEyeBlinkIndex.Keys.SelectMany(s => s.CloseAnimations).Select(a => a.Name).ToHashSet();
-        var openShapeNames = _advancedEyeBlinkIndex.Keys.SelectMany(s => s.OpenAnimations).Select(a => a.Name).ToHashSet();
-        var cancelerShapeNames = _advancedEyeBlinkIndex.Keys.SelectMany(s => s.CancelerBlendShapeNames).ToHashSet();
-
-        // 瞬き用のブレンドシェイプを複製する
-        // Trackingとの競合を避けるためと、他表情で使用されている場合にそれをリセットせず移動量の追加のみ行うようにするため。
-        var shapesToClone = new HashSet<string>(closeShapeNames);
-        shapesToClone.UnionWith(openShapeNames);
-        Action<Mesh, Mesh> onClone = (Mesh o, Mesh n) => ObjectRegistry.RegisterReplacedObject(o, n);
-        Action<string> onNotFound = (string name) => { Debug.LogError($"Shape not found: {name}"); };
-        var mapping = MeshHelper.CloneShapes(_sessionContext.FaceRenderer, shapesToClone, onClone, onNotFound, "_clone.blink");
-        foreach (var (settings, index) in _advancedEyeBlinkIndex.ToList())
-        {
-            var newSettings = settings.GetRenamed(mapping);
-            _advancedEyeBlinkIndex.Remove(settings);
-            _advancedEyeBlinkIndex[newSettings] = index;
-        }
-
-        var starePosition = animationGatePosition + new Vector3(PositionXStep, 0, 0);
-        foreach (var (advancedSettings, index) in _advancedEyeBlinkIndex.OrderBy(kvp => kvp.Value))
-        {
-            var stare = AddFTState(eyeBlinkLayer, $"Stare {index}", starePosition);
-            var close = AddFTState(eyeBlinkLayer, $"Close {index}", starePosition + new Vector3(PositionXStep, 0, 0));
-            var open = AddFTState(eyeBlinkLayer, $"Open {index}", starePosition + new Vector3(PositionXStep, 2 * PositionYStep, 0));
-
-            // AnimationGate -> Stare
-            var gateToStare = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-            gateToStare.SetDestination(stare);
-            // floatでequalsは使えないので2
-            if (index > 0)
-            {
-                gateToStare.Conditions = ImmutableList.Create(
-                    new AnimatorCondition()
-                    {
-                        parameter = BlinkAnimationModeAAP,
-                        mode = AnimatorConditionMode.Greater,
-                        threshold = VRCAAPHelper.IndexToValue(index - 1)
-                    }
-                );
-            }
-            if (index < 255)
-            {
-                gateToStare.Conditions = gateToStare.Conditions.Add(
-                    new AnimatorCondition()
-                    {
-                        parameter = BlinkAnimationModeAAP,
-                        mode = AnimatorConditionMode.Less,
-                        threshold = VRCAAPHelper.IndexToValue(index + 1)
-                    }
-                );
-            }
-            animationGate.Transitions = animationGate.Transitions.Add(gateToStare);
-
-            // Stare -> AnimationGate (OR)
-            var conditions = new List<AnimatorCondition>
-            {
-                new AnimatorCondition
-                {
-                    parameter = BlinkAnimationModeAAP,
-                    mode = AnimatorConditionMode.Greater,
-                    threshold = VRCAAPHelper.IndexToValue(index)
-                },
-                new AnimatorCondition
-                {
-                    parameter = BlinkAnimationModeAAP,
-                    mode = AnimatorConditionMode.Less,
-                    threshold = VRCAAPHelper.IndexToValue(index)
-                }
-            };
-            foreach (var condition in conditions)
-            {
-                var stareToGate = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-                stareToGate.SetDestination(animationGate);
-                stareToGate.Conditions = ImmutableList.Create(condition);
-                stare.Transitions = stare.Transitions.Add(stareToGate);
-            }
-
-            // Stare -> Close
-            var stareToClose = AnimatorHelper.CreateTransitionWithExitTime();
-            stareToClose.SetDestination(close);
-            stareToClose.Conditions = ImmutableList.Create(new AnimatorCondition()
-            {
-                parameter = BlinkAllowAAP,
-                mode = AnimatorConditionMode.Greater,
-                threshold = 0.99f
-            });
-            stare.Transitions = stare.Transitions.Add(stareToClose);
-
-            // Close -> Stare
-            var closeToStare = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-            closeToStare.SetDestination(stare);
-            closeToStare.Conditions = ImmutableList.Create(new AnimatorCondition()
-            {
-                parameter = BlinkAllowAAP,
-                mode = AnimatorConditionMode.Less,
-                threshold = 0.99f
-            });
-            close.Transitions = close.Transitions.Add(closeToStare);
-
-            // Close -> Open
-            var closeToOpen = AnimatorHelper.CreateTransitionWithExitTime();
-            closeToOpen.SetDestination(open);
-            close.Transitions = close.Transitions.Add(closeToOpen);
-
-            // Open -> Stare
-            var openToStare1 = AnimatorHelper.CreateTransitionWithExitTime();
-            openToStare1.SetDestination(stare);
-            open.Transitions = open.Transitions.Add(openToStare1);
-
-            var openToStare2 = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-            openToStare2.SetDestination(stare);
-            openToStare2.Conditions = ImmutableList.Create(new AnimatorCondition()
-            {
-                parameter = BlinkAllowAAP,
-                mode = AnimatorConditionMode.Less,
-                threshold = 0.99f
-            });
-            open.Transitions = open.Transitions.Add(openToStare2);
-
-            
-            // 瞬きの間隔を設定
-            if (!advancedSettings.UseRandomInterval)
-            {
-                var blinkDelayClip = AnimatorHelper.CreateDelayClip(advancedSettings.IntervalSeconds, $"BlinkDelay {index}");
-                stare.Motion = blinkDelayClip;
-            }
-            else
-            {
-                var blinkDelayClip = AnimatorHelper.CreateDelayClip(advancedSettings.RandomIntervalMinSeconds, $"BlinkDelay {index}");
-                stare.Motion = blinkDelayClip;
-
-                var maxMultiplier = 1f;
-                var minMultiplier = advancedSettings.RandomIntervalMinSeconds / advancedSettings.RandomIntervalMaxSeconds;
-                stare.EnsureBehavior<VRCAvatarParameterDriver>().parameters.Add(new VRC_AvatarParameterDriver.Parameter()
-                {
-                    type = VRC_AvatarParameterDriver.ChangeType.Random,
-                    name = BlinkDelayMultiplier,
-                    valueMin = minMultiplier,
-                    valueMax = maxMultiplier,
-                });
-                stare.SpeedParameter = BlinkDelayMultiplier;
-            }
-
-            // 目を閉じるアニメーションとAAPを設定
-            var closeClip = close.GetOrCreateClip($"Close {index}");
-            var closeCurve = new AnimationCurve();
-            closeCurve.AddKey(0f, 0f);
-            closeClip.SetFloatCurve("", typeof(Animator), BlinkBlinkingAAP, closeCurve);
-            AddAnimationToState(closeClip, advancedSettings.CloseAnimations.Select(a => a.ToGeneric(_sessionContext.BodyPath)));
-
-            // 目を開くアニメーションとAAPを設定
-            var openClip = open.GetOrCreateClip($"Open {index}");
-            var openCurve = new AnimationCurve();
-            openCurve.AddKey(0f, 1f);
-            openClip.SetFloatCurve("", typeof(Animator), BlinkBlinkingAAP, openCurve);
-            AddAnimationToState(openClip, advancedSettings.OpenAnimations.Select(a => a.ToGeneric(_sessionContext.BodyPath)));
-
-            starePosition.y += 3 * PositionYStep;
-        }
-
-        if (cancelerLayer is null) return;
-
-        var passThroughPosition = EntryStatePosition + new Vector3(PositionXStep, 0, 0);
-        var passThrough = AddFTState(cancelerLayer, "PassThrough", passThroughPosition);
-        AsPassThrough(passThrough);
-        var position = passThroughPosition + new Vector3(PositionXStep, 0, 0);
-        foreach (var (advancedSettings, index) in _advancedEyeBlinkIndex.OrderBy(kvp => kvp.Value))
-        {
-            if (!advancedSettings.IsCancelerEnabled()) continue;
-
-            var canceler = AddFTState(cancelerLayer, $"Canceler {index}", position);
-            var cancelerAnimation = advancedSettings.CancelerBlendShapeNames.Select(name => BlendShapeAnimation.SingleFrame(name, 0f).ToGeneric(_sessionContext.BodyPath));
-            AddAnimationToState(canceler, cancelerAnimation);
-
-            // passThrough -> canceler
-
-            var closeDuration = advancedSettings.CloseAnimations.Select(a => a.Time).Max();
-            var passThroughToCanceler = AnimatorHelper.CreateTransitionWithDurationSeconds(closeDuration);
-            passThroughToCanceler.SetDestination(canceler);
-            // cancelerが有効かつ、BlinkingがStare(-1f)を上回りClosing(0f)のとき有効
-            passThroughToCanceler.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = BlinkUseCancelerAAP,
-                    mode = AnimatorConditionMode.Greater,
-                    threshold = 0f
-                },
-                new AnimatorCondition()
-                {
-                    parameter = BlinkBlinkingAAP,
-                    mode = AnimatorConditionMode.Greater,
-                    threshold = -1f
-                },
-                new AnimatorCondition()
-                {
-                    parameter = BlinkBlinkingAAP,
-                    mode = AnimatorConditionMode.Less,
-                    threshold = 0.01f
-                }
-            );
-            passThrough.Transitions = ImmutableList.Create(passThroughToCanceler);
-
-            // canceler -> passThrough
-
-            // cancelerが無効な時は早期リターン
-            var cancelerToPassThrough1 = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-            cancelerToPassThrough1.SetDestination(passThrough);
-            cancelerToPassThrough1.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = BlinkUseCancelerAAP,
-                    mode = AnimatorConditionMode.Less,
-                    threshold = 1f
-                }
-            );
-            canceler.Transitions = canceler.Transitions.Add(cancelerToPassThrough1);
-
-            // Stare(-1f)の場合は早期リターン
-            var cancelerToPassThrough2 = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-            cancelerToPassThrough2.SetDestination(passThrough);
-            cancelerToPassThrough2.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = BlinkBlinkingAAP,
-                    mode = AnimatorConditionMode.Less,
-                    threshold = -0.99f
-                }
-            );
-            canceler.Transitions = canceler.Transitions.Add(cancelerToPassThrough2);
-
-            // Closing(0f)を上回り、Opening(1f)のとき逆方向にキャンセラーを適用する
-            var openDuration = advancedSettings.OpenAnimations.Select(a => a.Time).Max();
-            var cancelerToPassThrough3 = AnimatorHelper.CreateTransitionWithDurationSeconds(openDuration);
-            cancelerToPassThrough3.SetDestination(passThrough);
-            cancelerToPassThrough3.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = BlinkBlinkingAAP,
-                    mode = AnimatorConditionMode.Greater,
-                    threshold = 0.01f
-                }
-            );
-            canceler.Transitions = canceler.Transitions.Add(cancelerToPassThrough3);
-
-            position.y += 2 * PositionYStep;
-        }
-    }
-    
-    private void AddLipSyncLayer()
-    {
-        VirtualLayer? cancelerLayer = null;
-        VirtualLayer? lipSyncLayer = null;
-
-        if (_advancedLipSyncIndex.Any(kvp => kvp.Key.IsCancelerEnabled()))
-        {
-            cancelerLayer = AddFTLayer(_virtualController, "LipSync (Canceler)", LayerPriority);
-        }
-        lipSyncLayer = AddFTLayer(_virtualController, "LipSync", LayerPriority);
-        
-        var delayState = AddFTState(lipSyncLayer, "Delay", EntryStatePosition + new Vector3(-20, 2 * PositionYStep, 0));
-        var delayClip = AnimatorHelper.CreateDelayClip(0.1f);
-        delayState.Motion = delayClip;
-        
-        var enabledPosition = EntryStatePosition + new Vector3(PositionXStep, 0, 0);
-        var enabled = AddFTState(lipSyncLayer, "Enabled", enabledPosition);
-        var disabled = AddFTState(lipSyncLayer, "Disabled", enabledPosition + new Vector3(0, 2 * PositionYStep, 0));
-
-        _platformSupport.SetLipSyncTrack(enabled, true);
-        _platformSupport.SetLipSyncTrack(disabled, false);
-
-        var delayTransition = AnimatorHelper.CreateTransitionWithExitTime(1f, 0f);
-        delayTransition.SetDestination(enabled);
-        delayState.Transitions = ImmutableList.Create(delayTransition);
-
-        var enabledTransition = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-        enabledTransition.SetDestination(enabled);
-        enabledTransition.Conditions = ImmutableList.Create(new AnimatorCondition()
-        {
-            parameter = LipSyncAllowAAP,
-            mode = AnimatorConditionMode.Greater,
-            threshold = 0.99f
-        });
-        disabled.Transitions = ImmutableList.Create(enabledTransition);
-
-        var disabledTransition = AnimatorHelper.CreateTransitionWithDurationSeconds(0f);
-        disabledTransition.SetDestination(disabled);
-        disabledTransition.Conditions = ImmutableList.Create(new AnimatorCondition()
-        {
-            parameter = LipSyncAllowAAP,
-            mode = AnimatorConditionMode.Less,
-            threshold = 0.99f
-        });
-        enabled.Transitions = ImmutableList.Create(disabledTransition);
-
-        if (cancelerLayer is null) return;
-
-        var mutePosition = EntryStatePosition + new Vector3(PositionXStep, 0, 0);
-        var mute = AddFTState(cancelerLayer, "Mute", mutePosition);
-        mute.Motion = _emptyClip;
-        var position = mutePosition + new Vector3(PositionXStep, 0, 0);
-        var voiceParam = "Voice"; // Todo
-        _virtualController.EnsureParameterExists(AnimatorControllerParameterType.Float, voiceParam);
-        foreach (var (advancedSettings, index) in _advancedLipSyncIndex.OrderBy(kvp => kvp.Value))
-        {
-            if (!advancedSettings.IsCancelerEnabled()) continue;
-
-            var lipsyncing = AddFTState(cancelerLayer, $"Lipsyncing {index}", position);
-            var cancelerAnimation = advancedSettings.CancelerBlendShapeNames.Select(name => BlendShapeAnimation.SingleFrame(name, 0f).ToGeneric(_sessionContext.BodyPath));
-            AddAnimationToState(lipsyncing, cancelerAnimation);
-
-            // mute -> lipsyncing
-            var muteToLipsyncing = AnimatorHelper.CreateTransitionWithDurationSeconds(0.1f); // Todo: 設定可能にしても良いかも
-            muteToLipsyncing.SetDestination(lipsyncing);
-            muteToLipsyncing.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = LipSyncUseCancelerAAP,
-                    mode = AnimatorConditionMode.Greater,
-                    threshold = 0f
-                },
-                new AnimatorCondition()
-                {
-                    parameter = voiceParam,
-                    mode = AnimatorConditionMode.Greater,
-                    threshold = 0f
-                }
-            );
-            if (index > 0)
-            {   
-                muteToLipsyncing.Conditions = muteToLipsyncing.Conditions.Add(
-                    new AnimatorCondition()
-                    {
-                        parameter = LipSyncModeAAP,
-                        mode = AnimatorConditionMode.Greater,
-                        threshold = VRCAAPHelper.IndexToValue(index - 1)
-                    }
-                );
-            }
-            if (index < 255)
-            {
-                muteToLipsyncing.Conditions = muteToLipsyncing.Conditions.Add(
-                    new AnimatorCondition()
-                    {
-                        parameter = LipSyncModeAAP,
-                        mode = AnimatorConditionMode.Less,
-                        threshold = VRCAAPHelper.IndexToValue(index + 1)
-                    }
-                );
-            }
-            mute.Transitions = mute.Transitions.Add(muteToLipsyncing);
-
-            // lipsyncing -> mute
-            var lipsyncingToMute1 = AnimatorHelper.CreateTransitionWithDurationSeconds(0.1f);
-            lipsyncingToMute1.SetDestination(mute);
-            lipsyncingToMute1.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = voiceParam,
-                    mode = AnimatorConditionMode.Less,
-                    threshold = 0.01f
-                }
-            );
-            lipsyncing.Transitions = lipsyncing.Transitions.Add(lipsyncingToMute1);
-
-            var lipsyncingToMute2 = AnimatorHelper.CreateTransitionWithDurationSeconds(0.1f);
-            lipsyncingToMute2.SetDestination(mute);
-            lipsyncingToMute2.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = LipSyncUseCancelerAAP,
-                    mode = AnimatorConditionMode.Less,
-                    threshold = 0.01f
-                }
-            );
-            lipsyncing.Transitions = lipsyncing.Transitions.Add(lipsyncingToMute2);
-
-            var lipsyncingToMute3 = AnimatorHelper.CreateTransitionWithDurationSeconds(0.1f);
-            lipsyncingToMute3.SetDestination(mute);
-            lipsyncingToMute3.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = LipSyncModeAAP,
-                    mode = AnimatorConditionMode.Greater,
-                    threshold = VRCAAPHelper.IndexToValue(index)
-                }
-            );
-            lipsyncing.Transitions = lipsyncing.Transitions.Add(lipsyncingToMute3);
-
-            var lipsyncingToMute4 = AnimatorHelper.CreateTransitionWithDurationSeconds(0.1f);
-            lipsyncingToMute4.SetDestination(mute);
-            lipsyncingToMute4.Conditions = ImmutableList.Create(
-                new AnimatorCondition()
-                {
-                    parameter = LipSyncModeAAP,
-                    mode = AnimatorConditionMode.Less,
-                    threshold = VRCAAPHelper.IndexToValue(index)
-                }
-            );
-            lipsyncing.Transitions = lipsyncing.Transitions.Add(lipsyncingToMute4);
-            
-            position.y += PositionYStep;
-        }
+        _blinkInstaller.SetSettings(clip, facialSettings);
+        _lipSyncInstaller.SetSettings(clip, facialSettings);
     }
 }
