+++ Editor/Animator/InstallerBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ffbe486eb15dc024ca52441af734298a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
